+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:brick/pages/pages.dart';
-import 'package:brick/styles/styles.dart';
+import 'package:brick/src/app/pages/pages.dart';
+import 'package:brick/src/app/app.dart';
 import 'package:provider/provider.dart';
 
 void main() {
@@ -10,20 +10,4 @@ void main() {
     ],
     child: const App(),
   ));
-}
-
-class App extends StatelessWidget {
-  const App({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Brick',
-      // home: const Home(),
-      routes: routes,
-      theme: basicTheme,
-      debugShowCheckedModeBanner: false,
-    );
-  }
 }
