+++ lib/src/app/pages/splash/splash.dart
