+++ lib/src/app/styles/text.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+// Home BottomSheet TextField Struct Style
+const StrutStyle homeAddToDoStrutStyle = StrutStyle();
