--- lib/styles/styles.dart
@@ -1,2 +1,0 @@
-export './dimension.dart';
-export './theme.dart';
