--- lib/widgets/widgets.dart
@@ -1,5 +1,0 @@
-// Layouts
-export './layout/responsive_layout.dart';
-
-// ToDo
-export './todo/todo_tile.dart';
